refactor: move axdriver dyn part to axplat-dyn
